api/requirements.txt
azure-data-tables
azure-identity